Hacky pause rule implementation
The implementation of pausing as a rule requires access to the clock and controller instance.
A better implementation of pausing in the future would likely be as a game state.
I.e. the idea is that the game can be in different states (playing, paused, main menu, ...), implemented as the state pattern.