WIP implemented fast backtracking
Actually not very effective because of the enforced locality:
In the vast majority of cases, the last placed block is immediately adjacent to the unfillable block, so there is no
fast track to be taken, it's just a single step.
One might think that this just means that almost everything was good already anyways, and this change only fixes the
small amount of cases where locality is broken out of necessity, and an unfillable cell is encountered during that exact
move.
However, this 72x72 proves the opposite: it's not solved (before the heat death of the universe)